- SW Update:   - Updated docs with relevant info   - Finished validation of the loaded filesystem   - Added limits and integrity checks   - Improved filesystem struct structure (byte padding, ...)
# mvlach

+++ sw/src/consts/fat_flags.go
@@ -2,10 +2,27 @@
 package consts
 
 const (
-	// FatFree is the value of an unused FAT entry
-	FatFree int8 = -1
-	// FatFileEnd is the value of the last FAT entry of a file
-	FatFileEnd int8 = -2
-	// // FatBadCluster is the value of a bad cluster in the FAT - unused
-	// FatBadCluster int8 = -3
+	// FatFree is the value of an unused FAT entry.
+	//
+	// int32 is used to represent the FAT entries
+	// because the FAT entries can be negative and
+	// based on experiments and maximum file size,
+	// int32 is sufficient (maximum number of
+	// clusters is near 1 000 000)
+	FatFree int32 = -1
+	// FatFileEnd is the value of the last FAT entry of a file.
+	//
+	// int32 is used to represent the FAT entries
+	// because the FAT entries can be negative and
+	// based on experiments and maximum file size,
+	// int32 is sufficient (maximum number of
+	// clusters for 4GB FS is somewhere near 1 000 000).
+	FatFileEnd int32 = -2
+	// // FatBadCluster is the value of a bad cluster in the FAT.
+	// int32 is used to represent the FAT entries
+	// because the FAT entries can be negative and
+	// based on experiments and maximum file size,
+	// int32 is sufficient (maximum number of
+	// clusters for 4GB FS is somewhere near 1 000 000).		- not used
+	// FatBadCluster int32 = -3
 )
